Finish calculator project
I know, I got a bit too carried away and forgot to commit often. But
nothing drastic happened, so my first time has been delayed a bit.

Added functionality for the rest of the buttons to make the calculator
somewhat competent. Furthermore, checks have been added to ensure that
invalid inputs are rejected, as evident in the JS code, mostly to do
with not allowing the equals button to work unless there is a second
number, and which you can easily check in index.js.

CSS styling has been modified to ensure that special buttons, such as
operations and other useful actions, are highlighted and consequently
aid user experience.

Cosmetic CSS styling, such as box shadows and a slight blur filter, have
been added to make the calculator more appealing.

Thanks for the experience, Foundations course.